refactor: make chat clear and delete actions opt-in

+++ src/Livewire/Chat/Chat.php
@@ -310,6 +310,7 @@ class Chat extends Component
     {
         abort_unless(auth()->check(), 401);
         abort_unless($this->authParticipant !== null, 403, __('wirechat::chat.message_request.messages.accept_required'));
+        abort_unless($this->conversation->isSelf() || $this->conversation->isPrivate(), 403, 'This operation is not available for Groups.');
 
         // delete conversation
         $this->conversation->deleteFor($this->auth);
@@ -329,6 +330,7 @@ class Chat extends Component
     {
         abort_unless(auth()->check(), 401);
         abort_unless($this->authParticipant !== null, 403, __('wirechat::chat.message_request.messages.accept_required'));
+        abort_unless($this->conversation->isSelf() || $this->conversation->isPrivate(), 403, 'This operation is not available for Groups.');
 
         // delete conversation
         $this->conversation->clearFor($this->auth);
